#27 Refactoring
