+++ src/IM.Production.UI/src/contract/contract.component.html
@@ -20,7 +20,8 @@
                         <mat-grid-list cols="2" rowHeight="8:1">
 
                             <mat-grid-tile>Поставлено материала:</mat-grid-tile>
-                            <mat-grid-tile>{{contract?.totalCountCompleted | number}} / {{contract?.tillCount | number}} шт.</mat-grid-tile>
+                            <mat-grid-tile>{{contract?.totalCountCompleted | number}} / {{contract?.tillCount | number}}
+                                шт.</mat-grid-tile>
                             <mat-grid-tile>Общая сумма:</mat-grid-tile>
                             <mat-grid-tile>{{contract?.totalSumm | number}}</mat-grid-tile>
                             <mat-grid-tile>До даты:</mat-grid-tile>
