+++ src/IM.Production.UI/src/teams/teams.component.ts
@@ -2,7 +2,7 @@ import { MatDialog } from '@angular/material/dialog';
 import { Component, OnInit } from '@angular/core';
 import { TeamsService } from '../services/teams.service';
 import { NewTeamComponent } from '../new-team/new-team.component';
-import { TeamDto } from '../models/dtos/team';
+import { TeamDto } from '../models/dtos/team.dto';
 
 @Component({
     selector: 'app-teams',
